Do not run jobs on pull_request event on base repo
Currently, all jobs run doubled, onces because of the push event and one because of the pull_request event.

We want all jobs to run on push events, so that every commit is checked saparately, and we need the pull_request event to see jobs running on our own repo for pull requests from  forks.

To achieve this, we run jobs on pull_request events only if the owner of the action (us) matches the owner of the (HEAD) repository.

Signed-off-by: MichaIng <micha@dietpi.com>