ci: use Python versions matching Ubuntu versions
There is no point to test on different Ubuntu versions, when we use the same Python version on all of them. Use the Python version that is used on the respective Ubuntu version explicitly. The runners ship with the same minor version (3.8, 3.10 etc), but with more recent patch versions than provided by the Ubuntu APT repositories.

Signed-off-by: MichaIng <micha@dietpi.com>