ci: apply PAT for gh CLI
It is unclear why setting the GITHUB_TOKEN variable worked to grant the gh CLI permissions for opening a PR, since it documents that GH_TOKEN is needed. Probably it changed at some point. However, with the generic GITHUB_TOKEN, the PR does not trigger other workflows. "git push" triggers workflows on push event, but "gh pr" none on pull_request event, due to the GitHub limitation to prevent workflow loops.

Assign the correct variable for the gh CLI with the PAT applied for git as well, and do so only if gh is actually called.

Signed-off-by: MichaIng <micha@dietpi.com>