Fix test mail notification but with multiple addresses
For the test mail, the input string was never actually split into a list. It was applied as a single list entry instead.

Do the actual split by comma when passing the argument to the `send_mail()` function.

Actual notification mails do the splitting, even allowing semicolon, space and horizonal line splitting, stripping spaces and (resulting) empty list entries. But without breaking existing configs, for the test mail, I prefer to just be strict, assuring there are no unexpected characters, and splitting is done by comma.

Additionally, the text shown in case of a successful notification email test is updated to actually tell that it was a success. When multi-language support was added, replacing the copy&paste for this string was obviously forgotten. Weblate contributors, prepare for a new string.

Signed-off-by: MichaIng <micha@dietpi.com>